\name{ascrda-package}
\alias{ascrda-package}
\docType{package}

\title{Augmented SCRDA: improved rda classifier using nearest neighbor autocovariate}


\description{
The nearest neighbor autocovariates were defined by  Holmes and Adams (2003).
The algorithm in this package was developed in Islam (2008)
and the method was applied to the classification problems
in microarray experiments.
In this package improved RDA (regularized discriminant analysis) for class prediction 
in microarrays using nearest neighbour covariates is provided. 
Implementation of Support Vector Machine with feature selection is also available. 
A function for simulating microarray experiments is provided. 
}

\details{
\tabular{ll}{
Package: \tab ascrda\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2011-10-11\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}

\author{
M. S. Islam and A. I. McLeod
 <mislam22@uwo.ca>
}

\references{
Holmes, C. C. and Adams, N. M. (2003). 
Likelihood inference in nearest-neighbour classification models, 
Biometrika, 90(1), 99-112.
\url{http://biomet.oxfordjournals.org/cgi/content/abstract/90/1/99}

Islam, M. S. (2008). Periodicity, Change Detection and Prediction in Microarrays.
Ph.D. Thesis, The University of Western Ontario.
}


\keyword{ package }

